misc fixes

+++ app/src/main/java/com/lfsolutions/retail/model/dailysale/DailySaleRecord.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 
 data class DailySaleRecord(
-    @SerializedName("netTotal") var netTotal: Int? = null,
+    @SerializedName("netTotal") var netTotal: Double? = null,
     @SerializedName("dailySalesPaymentTermListDto") var dailySalesItem: ArrayList<DailySalesItem> = arrayListOf()
 )
