+++ app/src/main/java/com/lfsolutions/retail/model/sale/TaxForProduct.kt
@@ -8,7 +8,7 @@ data class TaxForProduct(
     @SerializedName("productId") var productId: Int? = null,
     @SerializedName("taxId") var taxId: Int? = null,
     @SerializedName("name") var name: String? = null,
-    @SerializedName("taxRate") var taxRate: Int? = null,
-    @SerializedName("taxValue") var taxValue: Int? = null
+    @SerializedName("taxRate") var taxRate: Double? = null,
+    @SerializedName("taxValue") var taxValue: Double? = null
 
 )
